Move map to its own collections overview component
